+++ _examples/demo.go
@@ -11,8 +11,8 @@ import (
 	"math"
 	"time"
 
-	ui "github.com/gizak/termui"
-	"github.com/gizak/termui/widgets"
+	ui "github.com/sqshq/termui"
+	"github.com/sqshq/termui/widgets"
 )
 
 func main() {