+++ _examples/gauge.go
@@ -10,8 +10,8 @@ import (
 	"fmt"
 	"log"
 
-	ui "github.com/gizak/termui"
-	"github.com/gizak/termui/widgets"
+	ui "github.com/sqshq/termui"
+	"github.com/sqshq/termui/widgets"
 )
 
 func main() {