+++ _examples/image.go
@@ -18,8 +18,8 @@ import (
 	"os"
 	"strings"
 
-	ui "github.com/gizak/termui"
-	"github.com/gizak/termui/widgets"
+	ui "github.com/sqshq/termui"
+	"github.com/sqshq/termui/widgets"
 )
 
 func main() {