+++ widgets/tabs.go
@@ -7,7 +7,7 @@ package widgets
 import (
 	"image"
 
-	. "github.com/gizak/termui"
+	. "github.com/sqshq/termui"
 )
 
 // TabPane is a renderable widget which can be used to conditionally render certain tabs/views.